python/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
numpy==1.24.3
networkx==3.1
scikit-image==0.21.0
segmentation-models-pytorch==0.3.3
pillow==9.5.0
tifffile==2023.4.12
scikit-learn==1.3.0
osmnx==1.3.0
requests==2.31.0
matplotlib==3.7.1
geopy 